email: make SES provider optional
